Added unpublishing test and moved singal processor
